+++ react/components/LoginComponent.js
@@ -15,6 +15,7 @@ import { truncateString } from '../utils/format-string'
 import { translate } from '../utils/translate'
 import { LoginPropTypes } from '../propTypes'
 import OneTapSignin from './OneTapSignin'
+import getBindingAddress from '../utils/getBindingAddress'
 
 import styles from '../styles.css'
 import Loading from './Loading'
@@ -58,6 +59,7 @@ class LoginComponent extends Component {
     } = this.props
 
     const pathname = history && history.location && history.location.pathname
+    const search = history && history.location && history.location.search
 
     const iconClasses = 'flex items-center'
     const iconLabel = iconLabelProfile || translate('store/login.signIn', intl)
@@ -85,7 +87,8 @@ class LoginComponent extends Component {
     if (loginButtonAsLink) {
       const linkTo = sessionProfile ? 'store.account' : 'store.login'
       const returnUrl =
-        !sessionProfile && `returnUrl=${encodeURIComponent(pathname)}`
+        !sessionProfile && `returnUrl=${encodeURIComponent(`${pathname}${search}`)}`
+      const bindingAddress = getBindingAddress()
       return (
         <div className={styles.buttonLink}>
           <ButtonWithIcon
@@ -96,7 +99,15 @@ class LoginComponent extends Component {
               )
             }
             iconPosition={showIconProfile ? 'left' : 'right'}
-            onClick={() => navigate({ page: linkTo, query: returnUrl })}
+            onClick={() => navigate({
+              page: linkTo,
+              query: new URLSearchParams({
+                returnUrl,
+                ...(bindingAddress && {
+                  bindingAddress,
+                }),
+              }).toString()
+            })}
           >
             {buttonContent}
           </ButtonWithIcon>