+++ react/utils/redirect.js
@@ -1,3 +1,6 @@
+import { ROOT_PATH } from '../common/global'
+import getBindingAddress from './getBindingAddress'
+
 export const getReturnUrl = () => {
   if (!window || !window.location) {
     return null
@@ -12,14 +15,18 @@ export const getDefaultRedirectUrl = isHeaderLogin => {
   if (isHeaderLogin) {
     return `${window.location.pathname}${window.location.search}`
   }
-  return '/'
+  return null
 }
 
 export const jsRedirect = ({ runtime, isHeaderLogin }) => {
   const url = getReturnUrl() || getDefaultRedirectUrl(isHeaderLogin)
 
   if (!url) {
-    return
+    const __bindingAddress = getBindingAddress()
+    const queryString = __bindingAddress
+      ? new URLSearchParams({ __bindingAddress }).toString()
+      : ''
+    return `${ROOT_PATH}/?${queryString}`
   }
 
   runtime.navigate({